token works here

+++ src/main/java/dev/danvega/jwt/record/User.java
@@ -4,7 +4,7 @@ import dev.danvega.jwt.entity.UserEntity;
 import lombok.Builder;
 
 @Builder
-public record User(Long id, String username, String password) {
+public record User(Integer id, String username, String password) {
 
     public static User from(UserEntity userEntity){
         return User.builder()
